lambda aws + Makefile
