+++ TODO.md
@@ -9,6 +9,7 @@ Creating docker image of an application
 Prepare AWS Lambda implementation and package preparation 
 
 Pack lambda 3rd party library dependencies into Lambda Layer 
+- pocwiczyc ze struktura folderow + 1 warstwa zamiast 2
 
 <br>
 
@@ -24,15 +25,27 @@ Configure API Gateway to use the Lambda (AWS console)
 2020-10-02 
 
 Using AWS services with boto3 (Secrets Manager) for storing DB password 
+- dodac - @lru_cache(maxsize=None)
+- osobna funkcja wolajaca klienta - get_secrets_manager_client()
 
-Mocking AWS services with moto and localstack 
+-dodac type annotations
 
+Mocking AWS services with moto (ut) and localstack (component tests) 
+- TODO
+
+Makefile scripts for building lambda packages & layers
+- TODO
 <br>
 
 2020-10-09 
 
+Authorization and authentication (API key, JWT) 
+
+<br>
+
+one day..
+<br>
 Application deployment in ECS using RDS 
 
 Terraform introduction - optional 
 
-Authorization and authentication (API key, JWT)
